requirements.txt
astroid==1.6.5
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
et-xmlfile==1.0.1
-e git+https://github.com/mobolic/facebook-sdk.git@667cc713084fb64d3249b0abddde3b02277b0a44#egg=facebook_sdk
google-api-core==1.2.1
google-auth==1.5.0
google-cloud-language==1.0.2
googleapis-common-protos==1.5.3
grpcio==1.12.1
idna==2.7
isort==4.3.4
jdcal==1.4
langdetect==1.0.7
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.14.5
openpyxl==2.5.4
pandas==0.23.1
pprint==0.1
protobuf==3.6.0
pyasn1==0.4.3
pyasn1-modules==0.2.1
pylint==1.9.2
python-dateutil==2.7.3
pytz==2018.4
requests==2.19.1
rope==0.10.7
rsa==3.4.2
six==1.11.0
urllib3==1.23
wrapt==1.10.11
yapf==0.22.0
